requirements.txt
# Telegram Bot
python-telegram-bot==20.7

# Async HTTP
aiohttp==3.9.1
httpx==0.26.0

# Reddit API
asyncpraw==7.7.1

# Twitter API
tweepy==4.14.0

# Google Trends
pytrends==4.9.2

# База данных
sqlalchemy==2.0.23
aiosqlite==0.19.0
alembic==1.13.1

# Redis для кэша и очередей
redis==5.0.1
aioredis==2.0.1

# Графики
matplotlib==3.8.2
plotly==5.18.0
kaleido==0.2.1

# Данные
pandas==2.1.4
numpy==1.26.2

# ML (опционально)
scikit-learn==1.3.2

# Утилиты
python-dotenv==1.0.0
pytz==2023.3
croniter==2.0.1

# Веб-скрейпинг
beautifulsoup4==4.12.2
lxml==5.0.0

# Валидация
pydantic==2.5.2